feat: compiler: start to write code generator
