+++ src/frontend/compiler/src/CodeGenerator.hs
@@ -2,37 +2,40 @@
 
 module CodeGenerator where
 
+
 import Ast
-import FFI
 import Instruction
+import FFIStructure
 
 import Control.Monad.State
+import Control.Monad.Except
+
 
 ----------------------------------------------------------
 -- instance of code generator
 ----------------------------------------------------------
-class CodeGenerator a where
-    -- state is         : (Ast, TableMetadata)
-    -- success return   : [Instruction]
-    -- fail return      : String
-    genCode :: StateT (a, TableMetadata, [Instruction]) (Either String) [Instruction]
+type CodeGenState = ([TableMetadata], [Instruction])
+
+type CodeGenEnv   = ExceptT String (State CodeGenState) [Instruction]
 
 
-instance CodeGenerator Expr where
-    genCode =
-        let
-            haveIndex = False
-        in do
-            (ast, metadata, instructions) <- get
-            if haveIndex
-            then notImplemented
-            else notImplemented
-        where
+cExpr :: Expr -> CodeGenEnv
+cExpr expr = do
+    (metadata, instructions) <- lift get
+    case expr of
+        BinExpr op e1 e2 -> cExpr e1 >> cExpr e2 >> cOp op >> get >>= put
+        _ -> put (metadata, instructions ++ [Instruction opColumn 0 0 ""])
+    snd <$> lift get
 
+
+cOp :: BinOp -> CodeGenEnv
+cOp op = do
+    (metadata, instructions) <- lift get
+    put (metadata, instructions ++ [Instruction opLt 0 0 ""])
+    snd <$> lift get
 
 
 ----------------------------------------------------------
 -- some help functions
 ----------------------------------------------------------
 notImplemented = error "Not Implemented Error."
-
